Additional code cleanup Moved classes into seperate files Moved a few extension methods around to consolidate
