+++ PackageExplorer/Models/NuGetFrameworkModel.cs
@@ -117,139 +117,4 @@ namespace NuGet.Frameworks
 			PortableFrameworks.Add(frameWork);
 		}
 	}
-
-	/// <summary>
-	/// Class NuGetTargetFrameworkItem.
-	/// </summary>
-	[PropertyChanged.ImplementPropertyChanged]
-	public class NuGetTargetFrameworkItem
-	{
-		/// <summary>
-		/// Gets or sets the name.
-		/// </summary>
-		/// <value>The name.</value>
-		public string Name { get; set; }
-
-		/// <summary>
-		/// Gets or sets the short name.
-		/// </summary>
-		/// <value>The short name.</value>
-		public string ShortName { get; set; }
-
-		/// <summary>
-		/// Gets or sets a value indicating whether this instance is default.
-		/// </summary>
-		/// <value><c>true</c> if this instance is default; otherwise, <c>false</c>.</value>
-		public bool IsDefault { get; set; }
-
-		/// <summary>
-		/// Gets or sets a value indicating whether this instance is selected.
-		/// </summary>
-		/// <value><c>true</c> if this instance is selected; otherwise, <c>false</c>.</value>
-		public bool IsSelected { get; set; }
-
-		/// <summary>
-		/// Gets or sets the selected item.
-		/// </summary>
-		/// <value>The selected item.</value>
-		public NuGetTargetFrameworkItem SelectedItem { get; set; }
-
-		/// <summary>
-		/// Gets or sets the framework.
-		/// </summary>
-		/// <value>The framework.</value>
-		public NuGetFramework Framework { get; set; }
-
-		/// <summary>
-		/// Gets or sets the items.
-		/// </summary>
-		/// <value>The items.</value>
-		public ObservableCollection<NuGetTargetFrameworkItem> Items { get; set; } = new ObservableCollection<NuGetTargetFrameworkItem>();
-
-		/// <summary>
-		/// Returns a <see cref="System.String" /> that represents this instance.
-		/// </summary>
-		/// <returns>A <see cref="System.String" /> that represents this instance.</returns>
-		public override string ToString()
-		{
-			return Name;
-		}
-	}
-
-	/// <summary>
-	/// Class NuGetTargetFrameworkItemExtensions.
-	/// </summary>
-	public static class NuGetTargetFrameworkItemExtensions
-	{
-		/// <summary>
-		/// Ases the targeted platform.
-		/// </summary>
-		/// <param name="items">The items.</param>
-		/// <returns>System.String.</returns>
-		public static string AsTargetedPlatformPath(this ObservableCollection<NuGetTargetFrameworkItem> items)
-		{
-			var selectedPlatforms = (from platform in items
-				where platform.IsSelected
-				select platform.SelectedItem?.Framework ?? platform?.Framework).ToList();
-
-			return selectedPlatforms.AsTargetedPlatformPath();
-		}
-
-		/// <summary>
-		/// Ases the targeted platform path.
-		/// </summary>
-		/// <param name="items">The items.</param>
-		/// <returns>System.String.</returns>
-		public static string AsTargetedPlatformPath(this IEnumerable<NuGetFramework> items)
-		{
-			if (!items.Any()) return string.Empty;
-
-			// Lets reduce the list of frameworks down to its base amount (remove all duplicate/equilivant frameworks)
-			var frameworkReducer = new FrameworkReducer();
-			var workingPlatforms = frameworkReducer.Reduce(items);
-			if (workingPlatforms == null || !workingPlatforms.Any()) workingPlatforms = items;
-
-			if (!workingPlatforms.Any())
-			{
-				return string.Empty;
-			}
-
-			var str = string.Join("+", workingPlatforms.Select(x => x.GetShortFolderName()));
-
-			if (!string.IsNullOrEmpty(str) && !str.Contains("portable-"))
-			{
-				str = "portable-" + str;
-			}
-
-			return str;
-		}
-
-		/// <summary>
-		/// Determines whether the list of NuGetFrameworks are valid or not.
-		/// </summary>
-		/// <param name="frameworks">The frameworks.</param>
-		/// <returns><c>true</c> if [is valid target platform] [the specified frameworks]; otherwise, <c>false</c>.</returns>
-		public static bool IsValidTargetPlatform(this IEnumerable<NuGetFramework> frameworks)
-		{
-			if (!frameworks.Any()) return false; // If we don't have any frameworks, its not valid
-			if (frameworks.Count() == 1) return true; // if we only have one framework, its always valid
-
-			// This is a work in progress to try to "validate" a target platform string before building it
-			var frameworkNameProvier = new FrameworkNameProvider(new[] {DefaultFrameworkMappings.Instance},
-				new[] {DefaultPortableFrameworkMappings.Instance});
-
-			int profileNumber;
-			if (!frameworkNameProvier.TryGetPortableProfile(frameworks, out profileNumber) && frameworks.Count() > 1)
-			{
-				return false; // not a valid combination
-			}
-
-			if (profileNumber != -1)
-			{
-				frameworkNameProvier.TryGetPortableFrameworks(profileNumber, out frameworks);
-			}
-
-			return frameworks.Any();
-		}
-	}
 }
